# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: github/chapool/go-wallet
  version: 0.1.0
paths:
...
  /api/v1/wallet/collect:
    post:
      security:
      - Bearer: []
      description: |-
        Manually trigger collection for a specific wallet.
        Moves funds from user wallet to hot wallet.
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - wallet
      summary: Trigger manual collection
      operationId: PostCollectRoute
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/postCollectPayload'
      responses:
        "200":
          description: Collection initiated successfully
          schema:
            $ref: '#/definitions/collectResponse'
        "400":
          description: PublicHTTPValidationError
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
  /api/v1/wallet/collects:
    get:
      security:
      - Bearer: []
      description: |-
        Get collection history for the authenticated user.
        Returns all collect transactions (type='collect') for user wallets.
      produces:
      - application/json
      tags:
      - wallet
      summary: Get collect history
      operationId: GetCollectsRoute
      parameters:
      - type: integer
        description: Chain ID
        name: chain_id
        in: query
      - enum:
        - confirmed
        - safe
        - finalized
        - failed
        type: string
        description: Filter by transaction status
        name: status
        in: query
      - minimum: 0
        type: integer
        default: 0
        description: Offset used for pagination, number of records to skip
        name: offset
        in: query
      - maximum: 500
        minimum: 1
        type: integer
        default: 50
        description: Limit used for pagination, number of records to retrieve
        name: limit
        in: query
      responses:
        "200":
          description: Collect history retrieved successfully
          schema:
            $ref: '#/definitions/getCollectsResponse'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
...
  /api/v1/wallet/withdraw/{withdrawId}/approve:
    post:
      security:
      - Bearer: []
      description: |-
        Approve a withdraw request and process it.
        Only admin users can approve withdraw requests.
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - wallet
      summary: Approve withdraw request (Admin only)
      operationId: PostApproveWithdrawRoute
      parameters:
      - type: string
        format: uuid
        description: Withdraw ID to approve
        name: withdrawId
        in: path
        required: true
      responses:
        "200":
          description: Withdraw approved and processed successfully
          schema:
            $ref: '#/definitions/withdrawResponse'
        "400":
          description: PublicHTTPValidationError
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "403":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "500":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
...
definitions:
...
  collectItem:
    type: object
    required:
    - id
    - chain_id
    - tx_hash
    - from_addr
    - to_addr
    - amount
    - status
    - block_no
    - created_at
    properties:
      amount:
        type: string
        example: "1.5"
      block_no:
        type: integer
        format: int64
        example: 12345678
      chain_id:
        type: integer
        example: 1
      confirmation_count:
        type: integer
        format: int64
        example: 5
      created_at:
        type: string
        format: date-time
      from_addr:
        type: string
        example: "0x742d35Cc6634C0532925a3b8D4C9db96C4b4d8b6"
      id:
        type: string
        format: uuid
      status:
        type: string
        enum:
        - confirmed
        - safe
        - finalized
        - failed
        example: confirmed
      to_addr:
        type: string
        example: 0x8ba1f109551bD432803012645Hac136c22C177E7
      tx_hash:
        type: string
        example: 0x...
  collectResponse:
    type: object
    required:
    - message
    - wallet_id
    properties:
      message:
        type: string
        example: Collection initiated successfully
      tx_hash:
        description: Transaction hash if collection was successful
        type: string
        example: "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
      wallet_id:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
...
  getCollectsResponse:
    type: object
    required:
    - collects
    properties:
      collects:
        type: array
        items:
          $ref: '#/definitions/collectItem'
...
  httpValidationErrorDetail:
    type: object
    required:
    - key
    - in
    - error
    properties:
      error:
        description: Error describing field validation failure
        type: string
      in:
        description: Indicates how the invalid field was provided
        type: string
      key:
        description: Key of field failing validation
        type: string
...
  postCollectPayload:
    type: object
    properties:
      chain_id:
        description: Chain ID (required if wallet_id is not provided, will use wallet's
          chain_id if wallet_id is provided)
        type: integer
        example: 1
      wallet_id:
        description: Wallet ID to collect from (optional, if not provided, chain_id
          must be provided)
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
...
  publicHttpError:
    type: object
    required:
    - status
    - type
    - title
    properties:
      detail:
        description: More detailed, human-readable, optional explanation of the error
        type: string
        example: User is lacking permission to access this resource
      status:
        description: HTTP status code returned for the error
        type: integer
        format: int64
        maximum: 599
        minimum: 100
        x-go-name: Code
        example: 403
      title:
        description: Short, human-readable description of the error
        type: string
        example: Forbidden
      type:
        $ref: '#/definitions/publicHttpErrorType'
  publicHttpErrorType:
    description: Type of error returned, should be used for client-side error handling
    type: string
    enum:
    - generic
    - PUSH_TOKEN_ALREADY_EXISTS
    - OLD_PUSH_TOKEN_NOT_FOUND
    - ZERO_FILE_SIZE
    - USER_DEACTIVATED
    - INVALID_PASSWORD
    - NOT_LOCAL_USER
    - TOKEN_NOT_FOUND
    - TOKEN_EXPIRED
    - USER_ALREADY_EXISTS
    - MALFORMED_TOKEN
    - LAST_AUTHENTICATED_AT_EXCEEDED
    - MISSING_SCOPES
  publicHttpValidationError:
    type: object
    required:
    - validationErrors
    allOf:
    - $ref: '#/definitions/publicHttpError'
    properties:
      validationErrors:
        description: List of errors received while validating payload against schema
        type: array
        items:
          $ref: '#/definitions/httpValidationErrorDetail'
...
  withdrawItem:
    type: object
    required:
    - id
    - user_id
    - to_address
    - token_id
    - amount
    - status
    - created_at
    properties:
      amount:
        type: string
        example: "1.5"
      created_at:
        type: string
        format: date-time
      fee:
        type: string
        example: "0.001"
      id:
        type: string
        format: uuid
      status:
        type: string
        enum:
        - user_withdraw_request
        - signing
        - pending
        - processing
        - confirmed
        - failed
        example: user_withdraw_request
      to_address:
        type: string
        example: "0x742d35Cc6634C0532925a3b8D4C9db96C4b4d8b6"
      token_id:
        type: integer
        example: 1
      tx_hash:
        type: string
        example: 0x...
      user_id:
        type: string
        format: uuid
  withdrawResponse:
    type: object
    required:
    - withdraw
    properties:
      withdraw:
        $ref: '#/definitions/withdrawItem'
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query