# internal/types/withdraw_item.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"
	"encoding/json"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// WithdrawItem withdraw item
//
// swagger:model withdrawItem
type WithdrawItem struct {

	// amount
	// Example: 1.5
	// Required: true
	Amount *string `json:"amount"`

	// created at
	// Required: true
	// Format: date-time
	CreatedAt *strfmt.DateTime `json:"created_at"`

	// fee
	// Example: 0.001
	Fee string `json:"fee,omitempty"`

	// id
	// Required: true
	// Format: uuid
	ID *strfmt.UUID `json:"id"`

	// status
	// Example: user_withdraw_request
	// Required: true
	// Enum: [user_withdraw_request signing pending processing confirmed failed]
	Status *string `json:"status"`

	// to address
	// Example: 0x742d35Cc6634C0532925a3b8D4C9db96C4b4d8b6
	// Required: true
	ToAddress *string `json:"to_address"`

	// token id
	// Example: 1
	// Required: true
	TokenID *int64 `json:"token_id"`

	// tx hash
	// Example: 0x...
	TxHash string `json:"tx_hash,omitempty"`

	// user id
	// Required: true
	// Format: uuid
	UserID *strfmt.UUID `json:"user_id"`
}

// Validate validates this withdraw item
func (m *WithdrawItem) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateAmount(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateCreatedAt(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateStatus(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateToAddress(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateTokenID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateUserID(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *WithdrawItem) validateAmount(formats strfmt.Registry) error {

	if err := validate.Required("amount", "body", m.Amount); err != nil {
		return err
	}

	return nil
}

func (m *WithdrawItem) validateCreatedAt(formats strfmt.Registry) error {

	if err := validate.Required("created_at", "body", m.CreatedAt); err != nil {
		return err
	}

	if err := validate.FormatOf("created_at", "body", "date-time", m.CreatedAt.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *WithdrawItem) validateID(formats strfmt.Registry) error {

	if err := validate.Required("id", "body", m.ID); err != nil {
		return err
	}

	if err := validate.FormatOf("id", "body", "uuid", m.ID.String(), formats); err != nil {
		return err
	}

	return nil
}

var withdrawItemTypeStatusPropEnum []interface{}

func init() {
	var res []string
	if err := json.Unmarshal([]byte(`["user_withdraw_request","signing","pending","processing","confirmed","failed"]`), &res); err != nil {
		panic(err)
	}
	for _, v := range res {
		withdrawItemTypeStatusPropEnum = append(withdrawItemTypeStatusPropEnum, v)
	}
}

const (

	// WithdrawItemStatusUserWithdrawRequest captures enum value "user_withdraw_request"
	WithdrawItemStatusUserWithdrawRequest string = "user_withdraw_request"

	// WithdrawItemStatusSigning captures enum value "signing"
	WithdrawItemStatusSigning string = "signing"

	// WithdrawItemStatusPending captures enum value "pending"
	WithdrawItemStatusPending string = "pending"

	// WithdrawItemStatusProcessing captures enum value "processing"
	WithdrawItemStatusProcessing string = "processing"

	// WithdrawItemStatusConfirmed captures enum value "confirmed"
	WithdrawItemStatusConfirmed string = "confirmed"

	// WithdrawItemStatusFailed captures enum value "failed"
	WithdrawItemStatusFailed string = "failed"
)

// prop value enum
func (m *WithdrawItem) validateStatusEnum(path, location string, value string) error {
	if err := validate.EnumCase(path, location, value, withdrawItemTypeStatusPropEnum, true); err != nil {
		return err
	}
	return nil
}

func (m *WithdrawItem) validateStatus(formats strfmt.Registry) error {

	if err := validate.Required("status", "body", m.Status); err != nil {
		return err
	}

	// value enum
	if err := m.validateStatusEnum("status", "body", *m.Status); err != nil {
		return err
	}

	return nil
}

func (m *WithdrawItem) validateToAddress(formats strfmt.Registry) error {

	if err := validate.Required("to_address", "body", m.ToAddress); err != nil {
		return err
	}

	return nil
}

func (m *WithdrawItem) validateTokenID(formats strfmt.Registry) error {

	if err := validate.Required("token_id", "body", m.TokenID); err != nil {
		return err
	}

	return nil
}

func (m *WithdrawItem) validateUserID(formats strfmt.Registry) error {

	if err := validate.Required("user_id", "body", m.UserID); err != nil {
		return err
	}

	if err := validate.FormatOf("user_id", "body", "uuid", m.UserID.String(), formats); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this withdraw item based on context it is used
func (m *WithdrawItem) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *WithdrawItem) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *WithdrawItem) UnmarshalBinary(b []byte) error {
	var res WithdrawItem
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}